+++ README.md
@@ -1,21 +1,33 @@
 # Benchmarks
 
-## IList vs List
+## IList vs List & Yield vs NoYield
 
 ```
-| Method | Count |         Mean |         Error |        StdDev | Ratio | RatioSD | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
-|------- |------ |-------------:|--------------:|--------------:|------:|--------:|-----:|------------:|------------:|------------:|--------------------:|
-|  IList |     1 |     30.52 ns |     0.6411 ns |     1.5846 ns |  1.00 |    0.00 |    2 |      0.0095 |           - |           - |                40 B |
-|   List |     1 |     15.44 ns |     0.3352 ns |     0.3292 ns |  0.49 |    0.03 |    1 |           - |           - |           - |                   - |
-|        |       |              |               |               |       |         |      |             |             |             |                     |
-|  IList |    10 |    115.49 ns |     2.3221 ns |     3.6832 ns |  1.00 |    0.00 |    2 |      0.0093 |           - |           - |                40 B |
-|   List |    10 |     39.36 ns |     0.8454 ns |     1.0064 ns |  0.34 |    0.01 |    1 |           - |           - |           - |                   - |
-|        |       |              |               |               |       |         |      |             |             |             |                     |
-|  IList |   100 |    814.72 ns |    15.3247 ns |    14.3347 ns |  1.00 |    0.00 |    2 |      0.0086 |           - |           - |                40 B |
-|   List |   100 |    304.37 ns |     5.9634 ns |     9.1068 ns |  0.37 |    0.01 |    1 |           - |           - |           - |                   - |
-|        |       |              |               |               |       |         |      |             |             |             |                     |
-|  IList | 10000 | 74,189.87 ns | 1,461.8007 ns | 1,795.2225 ns |  1.00 |    0.00 |    2 |           - |           - |           - |                40 B |
-|   List | 10000 | 29,060.01 ns |   606.4166 ns | 1,464.5641 ns |  0.39 |    0.02 |    1 |           - |           - |           - |                   - |
+  Method | Count |                Mean |             Error |            StdDev | Ratio | RatioSD | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
+-------- |------ |--------------------:|------------------:|------------------:|------:|--------:|-----:|------------:|------------:|------------:|--------------------:|
+   IList |     1 |            31.71 ns |         0.6744 ns |         1.9241 ns |  1.00 |    0.00 |    2 |      0.0095 |           - |           - |                40 B |
+    List |     1 |            16.26 ns |         0.3663 ns |         0.8116 ns |  0.51 |    0.04 |    1 |           - |           - |           - |                   - |
+         |       |                     |                   |                   |       |         |      |             |             |             |                     |
+   Yield |     1 |     1,990,015.40 ns |     5,264.2317 ns |     4,666.6042 ns |  1.00 |    0.00 |    1 |           - |           - |           - |                40 B |
+ NoYield |     1 |     1,991,413.52 ns |     4,474.6337 ns |     4,185.5751 ns |  1.00 |    0.00 |    1 |           - |           - |           - |                64 B |
+         |       |                     |                   |                   |       |         |      |             |             |             |                     |
+   IList |    10 |           127.18 ns |         2.5794 ns |         5.8220 ns |  1.00 |    0.00 |    2 |      0.0093 |           - |           - |                40 B |
+    List |    10 |            42.83 ns |         1.0150 ns |         2.9123 ns |  0.33 |    0.03 |    1 |           - |           - |           - |                   - |
+         |       |                     |                   |                   |       |         |      |             |             |             |                     |
+   Yield |    10 |    11,950,059.32 ns |    27,166.2902 ns |    25,411.3647 ns |  1.00 |    0.00 |    1 |           - |           - |           - |                40 B |
+ NoYield |    10 |    19,887,310.05 ns |   103,146.2087 ns |    96,483.0276 ns |  1.66 |    0.01 |    2 |           - |           - |           - |                96 B |
+         |       |                     |                   |                   |       |         |      |             |             |             |                     |
+   IList |   100 |           937.40 ns |        18.5009 ns |        45.7296 ns |  1.00 |    0.00 |    2 |      0.0086 |           - |           - |                40 B |
+    List |   100 |           334.12 ns |         8.2474 ns |        24.1881 ns |  0.36 |    0.03 |    1 |           - |           - |           - |                   - |
+         |       |                     |                   |                   |       |         |      |             |             |             |                     |
+   Yield |   100 |   101,449,942.67 ns |   465,846.7274 ns |   435,753.3178 ns |  1.00 |    0.00 |    1 |           - |           - |           - |                40 B |
+ NoYield |   100 |   199,264,617.78 ns |   474,979.2256 ns |   444,295.8624 ns |  1.96 |    0.01 |    2 |           - |           - |           - |               456 B |
+         |       |                     |                   |                   |       |         |      |             |             |             |                     |
+   IList |  1000 |         8,881.32 ns |       175.8008 ns |       385.8871 ns |  1.00 |    0.00 |    2 |           - |           - |           - |                40 B |
+    List |  1000 |         2,904.50 ns |        57.9653 ns |       117.0927 ns |  0.33 |    0.02 |    1 |           - |           - |           - |                   - |
+         |       |                     |                   |                   |       |         |      |             |             |             |                     |
+   Yield |  1000 |   997,205,526.67 ns | 2,155,423.1771 ns | 2,016,184.1772 ns |  1.00 |    0.00 |    1 |           - |           - |           - |                40 B |
+ NoYield |  1000 | 1,994,124,633.33 ns | 1,744,351.0692 ns | 1,631,667.0724 ns |  2.00 |    0.00 |    2 |           - |           - |           - |              4056 B |
 ```
 
 ## MessagePack-Csharp vs Protobuf-net vs Newtonsoft.Json
